fix(ext/ai): moving types to `global.d.ts`
