+++ ext/ai/js/onnxruntime/inference_api.js
@@ -1,90 +1,90 @@
 import { InferenceSession, Tensor } from 'ext:ai/onnxruntime/onnx.js';
 
 const DEFAULT_HUGGING_FACE_OPTIONS = {
-  hostname: 'https://huggingface.co',
-  path: {
-    template: '{REPO_ID}/resolve/{REVISION}/onnx/{MODEL_FILE}?donwload=true',
-    revision: 'main',
-    modelFile: 'model_quantized.onnx',
-  },
+    hostname: 'https://huggingface.co',
+    path: {
+        template: '{REPO_ID}/resolve/{REVISION}/onnx/{MODEL_FILE}?donwload=true',
+        revision: 'main',
+        modelFile: 'model_quantized.onnx',
+    },
 };
 
 /**
  * An user friendly API for onnx backend
  */
 class UserInferenceSession {
-  inner;
+    inner;
 
-  id;
-  inputs;
-  outputs;
+    id;
+    inputs;
+    outputs;
 
-  constructor(session) {
-    this.inner = session;
+    constructor(session) {
+        this.inner = session;
 
-    this.id = session.sessionId;
-    this.inputs = session.inputNames;
-    this.outputs = session.outputNames;
-  }
-
-  static async fromUrl(modelUrl) {
-    if (modelUrl instanceof URL) {
-      modelUrl = modelUrl.toString();
+        this.id = session.sessionId;
+        this.inputs = session.inputNames;
+        this.outputs = session.outputNames;
     }
 
-    const encoder = new TextEncoder();
-    const modelUrlBuffer = encoder.encode(modelUrl);
-    const session = await InferenceSession.fromBuffer(modelUrlBuffer);
+    static async fromUrl(modelUrl) {
+        if (modelUrl instanceof URL) {
+            modelUrl = modelUrl.toString();
+        }
 
-    return new UserInferenceSession(session);
-  }
+        const encoder = new TextEncoder();
+        const modelUrlBuffer = encoder.encode(modelUrl);
+        const session = await InferenceSession.fromBuffer(modelUrlBuffer);
 
-  static async fromHuggingFace(repoId, opts = {}) {
-    const hostname = opts?.hostname ?? DEFAULT_HUGGING_FACE_OPTIONS.hostname;
-    const pathOpts = {
-      ...DEFAULT_HUGGING_FACE_OPTIONS.path,
-      ...opts?.path,
-    };
-
-    const modelPath = pathOpts.template
-      .replaceAll('{REPO_ID}', repoId)
-      .replaceAll('{REVISION}', pathOpts.revision)
-      .replaceAll('{MODEL_FILE}', pathOpts.modelFile);
-
-    if (!URL.canParse(modelPath, hostname)) {
-      throw Error(`[Invalid URL] Couldn't parse the model path: "${modelPath}"`);
+        return new UserInferenceSession(session);
     }
 
-    return await UserInferenceSession.fromUrl(new URL(modelPath, hostname));
-  }
+    static async fromHuggingFace(repoId, opts = {}) {
+        const hostname = opts?.hostname ?? DEFAULT_HUGGING_FACE_OPTIONS.hostname;
+        const pathOpts = {
+            ...DEFAULT_HUGGING_FACE_OPTIONS.path,
+            ...opts?.path,
+        };
 
-  async run(inputs) {
-    const outputs = await core.ops.op_sb_ai_ort_run_session(this.id, inputs);
+        const modelPath = pathOpts.template
+            .replaceAll('{REPO_ID}', repoId)
+            .replaceAll('{REVISION}', pathOpts.revision)
+            .replaceAll('{MODEL_FILE}', pathOpts.modelFile);
 
-    // Parse to Tensor
-    for (const key in outputs) {
-      if (Object.hasOwn(outputs, key)) {
-        const { type, data, dims } = outputs[key];
+        if (!URL.canParse(modelPath, hostname)) {
+            throw Error(`[Invalid URL] Couldn't parse the model path: "${modelPath}"`);
+        }
 
-        outputs[key] = new UserTensor(type, data.buffer, dims);
-      }
+        return await UserInferenceSession.fromUrl(new URL(modelPath, hostname));
     }
 
-    return outputs;
-  }
+    async run(inputs) {
+        const outputs = await core.ops.op_sb_ai_ort_run_session(this.id, inputs);
+
+        // Parse to Tensor
+        for (const key in outputs) {
+            if (Object.hasOwn(outputs, key)) {
+                const { type, data, dims } = outputs[key];
+
+                outputs[key] = new UserTensor(type, data.buffer, dims);
+            }
+        }
+
+        return outputs;
+    }
 }
 
 class UserTensor extends Tensor {
-  constructor(type, data, dim) {
-    super(type, data, dim);
-  }
+    constructor(type, data, dim) {
+        super(type, data, dim);
+    }
 
-  async tryEncodeAudio(sampleRate) {
-    return await core.ops.op_sb_ai_ort_encode_tensor_audio(this.data, sampleRate);
-  }
+    async tryEncodeAudio(sampleRate) {
+        return await core.ops.op_sb_ai_ort_encode_tensor_audio(this.data, sampleRate);
+    }
 }
 
 export default {
-  RawSession: UserInferenceSession,
-  Tensor: UserTensor,
+    RawSession: UserInferenceSession,
+    RawTensor: UserTensor,
 };
